feat(bankofsimba): balance formatted

+++ week-07/day-2/bankofsimba/src/main/java/com/greenfoxacademy/bankofsimba/models/BankAccount.java
@@ -1,12 +1,14 @@
 package com.greenfoxacademy.bankofsimba.models;
 
+import java.text.DecimalFormat;
+
 public class BankAccount {
 
     private String name;
-    private int balance;
+    private double balance;
     private String animalType;
 
-    public BankAccount(String name, int balance, String animalType) {
+    public BankAccount(String name, double balance, String animalType) {
         this.name = name;
         this.balance = balance;
         this.animalType = animalType;
@@ -20,7 +22,7 @@ public class BankAccount {
         this.name = name;
     }
 
-    public int getBalance() {
+    public double getBalance() {
         return balance;
     }
